Main HTML page in progmem, loading external resources
The plan is to host JS / CSS off-board on a separate webserver.  This
works well with libraries like Vue, which allow us to inject an
application into an element on the page.  By doing so, we can keep the
on-board HTML quite small.

For now, of course, these static resources do very little except make
changes that allow us to test that they were successfully loaded.