Starting NTP integration, separate wifi indicator light
The goal is to wire up the Quiic relay on pin 5, which also controls the
on-board LED.  During development, this allows us to see what decision
the thermostat would take; after installation, it provides visual
feedback as to the heating state of the device.

To help do that, I've added an LED on pin 4 - the wifi indicator light,
which is used to show whether wifi is connected, connecting, or in an
error state.