Return ISO 8601 timestamp from REST API
I don't necessarily *need* to return this from the REST API, but it
does help verify that the device is fetching the correct localized
timestamp.

This makes use of three libraries:

- `NTPClient` for periodically syncing time from `pool.ntp.org`;
- `Timezone` to hardcode the local timezone and any DST transitions;
- `Time` to get year, month, etc. from the current timestamp.

Unfortunately, these libraries overlap somewhat in their functionality,
meaning that I have to pick and choose pieces of them to make this
work.

Right now the timezone is hardcoded.  We *could* instead compare the
browser timezone offset to the current on-board offset, and offer the
user the option to update the on-board offset if there's a mismatch -
but for now that's more work than it's worth.  (Note that this is
essentially what Google Calendar does when you use it from a different
timezone.)